+++ src/index.js
@@ -1,7 +1,7 @@
 exports.collar = require('./collar');
 exports.deepFreeze = require('./deepFreeze');
-exports.deepSeal = require('./deepSeal');
 exports.deepPreventExtensions = require('./deepPreventExtensions');
+exports.deepSeal = require('./deepSeal');
 exports.promisify = require('./promisify');
 exports.singular = require('./singular');
 exports.sort = require('./sort');