+++ src/internals/assert.js
@@ -1,0 +1,7 @@
+// Public
+
+module.exports = function(boolExpr, message) {
+  if (!boolExpr) {
+    throw new TypeError(message);
+  }
+};