+++ src/internals/isPlainObject.js
@@ -1,0 +1,4 @@
+// Public
+
+module.exports = (testVar) => !!(testVar && typeof testVar === 'object' &&
+    Object.prototype.toString.call(testVar) === '[object Object]');