+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## [2.0.0] - 2017-07-22
+### Changed
+- Removed assertions of all inputs from toEnum
+- toEnum does not mutate input anymore
+- Added option to deep (freeze, seal, preventExtensions) object prototype
+- promisify.all don't accept mutate anymore and always mutate module (breaking change)
+- promisify.all added option to promisify prototype chain
+- removed documentation from code
+
 ## [1.4.0] - 2017-07-19
 ### Added
 - deepPreventExtensions